Sign & Number column now use background color
A significant change, the sign and number columns are no longer styled
differently to the content are. Rather they both now use the background
colour. Line numbers themselves are now styled in a subdued grey color
so as to not be distracting.

Reason for the change?

I now use GitGutter, so the width of the sign and number column often
ends up being 4 characters wide. Such a wide block looks visually ugly
with the different color as previously used by this theme.

I had a look at: Gruvbox, Vim One, Nova and even VSCode and they all use
the background color for signs/number columns. I like that look and have
now adopted it with this theme.