fix(lualine): add missing modes for `terminal`
When using the `use_mode_colors = true` option for the window/tab components in lualine, the sections `b` and `y` may not display the correct colors in terminal mode due to the absence of color definitions within the theme script. Resulting in the highlight group, e.g. `lualine_b_command` being initialized with default (cleared) value.

While this issue might be related to how lualine handles highlight groups, the simplest solution is to add the missing mode configurations directly in the theme.

This simply copies the insert mode table to a new table called `terminal` with the same color values.